chore: refactor recovery methods
