--- packages/transaction-recovery/src/manual-relay-to-dest-chain/lib/recovery.spec.ts
@@ -1,3 +1,0 @@
-describe("recovery", () => {
-  test("should call recoverEvmToEvm tx when source chain and destination chain are EVM chains", async () => {});
-});
