+++ packages/transaction-recovery/src/manual-relay-to-dest-chain/lib/strategies/index.ts
@@ -1,0 +1,5 @@
+export * from "./recoverEvmToEvm";
+export * from "./recoverEvmToIbc";
+export * from "./recoverIbcToEvm";
+export * from "./recoverIbcToIbc";
+export * from "./types";
