+++ packages/transaction-recovery/src/manual-relay-to-dest-chain/lib/strategies/recoverIbcToIbc.ts
@@ -1,0 +1,10 @@
+import type { RecoveryDependencies, RecoveryParams } from "..";
+import { sendAxelarRouteMessageTx } from "../AxelarRouteMessageTx";
+
+export async function recoverIbcToIbc(
+  params: RecoveryParams,
+  deps: RecoveryDependencies
+) {
+  const routeMessageTxResponse = await sendAxelarRouteMessageTx(params, deps);
+  return [routeMessageTxResponse];
+}
