+++ packages/transaction-recovery/src/manual-relay-to-dest-chain/lib/strategies/types.ts
@@ -1,0 +1,16 @@
+import type { SearchGMPResponseData } from "@axelarjs/api";
+import type { AxelarQueryClientService } from "@axelarjs/cosmos";
+
+import type { ManualRelayToDestChainDependencies } from "../../isomorphic";
+import { type ChainConfig } from "../helper";
+
+export type RecoveryDependencies = ManualRelayToDestChainDependencies & {
+  axelarQueryRpcClient: AxelarQueryClientService;
+};
+
+export type RecoveryParams = {
+  searchGMPData: SearchGMPResponseData;
+  srcChainConfig: ChainConfig;
+  destChainConfig: ChainConfig;
+  escapeAfterConfirm?: boolean | undefined;
+};
